Rename all instances of Duke to SadCat Update README docs and replace current Ui.png image.
