+++ src/main/java/sadcat/exceptions/SadCatException.java
