+++ src/test/java/sadcat/SadCatTest.java
@@ -1,4 +1,4 @@
-package duke;
+package sadcat;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
